requirements.txt
firebase==4.0.1
getmac==0.9.2
ifcfg==0.23
pandas==1.5.3
Pillow==9.4.0
Pillow==9.5.0
pyotp==2.8.0
Pyrebase4==4.6.0
python_nmap==0.7.1
qrcode==7.4.2
Requests==2.30.0
scapy==2.5.0
streamlit==1.20.0
streamlit_extras==0.2.7
streamlit_lottie==0.0.3
